Restruct

+++ SuperAgent/device/app/Script/Global/Events.js
@@ -47,12 +47,20 @@ function OnWorkflowStart(name) {
 	if (name == "Visit")
 	{
 
-		var outlet = GlobalWorkflow.GetOutlet();
+		var outlet = $.workflow.outlet;
 
-		CreateQuestionnareTable(outlet);
-		CreateQuestionsTable(outlet);
-		CreateSKUQuestionsTable(outlet);
-		SetSteps(outlet);
+	//	CreateQuestionnareTable(outlet);
+  //	CreateQuestionsTable(outlet);
+	//	CreateSKUQuestionsTable(outlet);
+
+		var q = new Query("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='USR_WorkflowSteps'");
+		var check = q.ExecuteScalar();
+
+		if (parseInt(check) == parseInt(0)) {
+			var q = new Query("CREATE TABLE " + " USR_WorkflowSteps (StepOrder, Skip, Value, NextStep, LastStep)");
+			q.Execute();
+		}
+
 	}
 
 }
@@ -725,12 +733,24 @@ function DeleteFromList(item, collection) {
 }
 
 function ClearUSRTables(){
-	var q = new Query("DELETE FROM USR_Questionnaires");
-	q.Execute();
+	var q = new Query("SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name ='USR_Questionnaires'");
+	var TableQuestion = q.ExecuteScalar();
+	if (parseInt(TableQuestion)==parseInt(1)) {
+		var q = new Query("DELETE FROM USR_Questionnaires");
+		q.Execute();
+	}
 
-	var q = new Query("DELETE FROM USR_Questions");
-	q.Execute();
+	var q = new Query("SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name ='USR_Questions'");
+	var TableQuestion = q.ExecuteScalar();
+	if (parseInt(TableQuestion)==parseInt(1)) {
+		var q = new Query("DELETE FROM USR_Questions");
+		q.Execute();
+	}
 
-	var q = new Query("DELETE FROM USR_SKUQuestions");
-	q.Execute();
+	var q = new Query("SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name ='USR_SKUQuestions'");
+	var TableQuestion = q.ExecuteScalar();
+	if (parseInt(TableQuestion)==parseInt(1)) {
+		var q = new Query("DELETE FROM USR_SKUQuestions");
+		q.Execute();
+	}
 }
